backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-openai
chromadb
# ---- Core backend ----
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

# ---- LLM client (OpenAI-compatible; used for Groq) ----
openai==1.40.6
httpx==0.27.2

# ---- RAG / Vector DB ----
chromadb==0.5.5
sentence-transformers==3.0.1

# ---- Document ingestion ----
pypdf==4.3.1

#ChromaDB
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1

python-multipart==0.0.9

aiosqlite==0.20.0
